split vector maths in own file
